requirements.txt
torch
transformers
pillow
numpy
requests